fix: Adapt tests to new tsconfig options

+++ backend/test/integration/User.test.ts
@@ -8,11 +8,15 @@ import LocalVMProvider from "../../src/providers/LocalVMProvider";
 import { beforeAll, describe, it, expect, afterAll } from "@jest/globals";
 
 const app = express();
-let connection: MongoClient = null;
-let instance: MongoDBPersister = null;
+let connection: MongoClient;
+let instance: MongoDBPersister;
 
 beforeAll(async () => {
-  connection = await MongoClient.connect(process.env.MONGO_URL);
+  const ENV_MONGO_URL = process.env.MONGO_URL;
+
+  if (!ENV_MONGO_URL) throw new Error("MongoDB url not provided (MONGO_URL).");
+
+  connection = await MongoClient.connect(ENV_MONGO_URL);
   try {
     await connection.db().dropCollection("users");
   } catch (err) {}
@@ -22,13 +26,13 @@ beforeAll(async () => {
     password: "testpassword",
     environments: [],
   });
-  instance = new MongoDBPersister(process.env.MONGO_URL);
+  instance = new MongoDBPersister(ENV_MONGO_URL);
   app.use(
     APIRoutes(
       instance,
       [new MongoDBAuthenticationProvider(instance)],
-      new LocalVMProvider()
-    )
+      new LocalVMProvider(),
+    ),
   );
 });
